Update tooltip name, prop, behavior on click
