+++ packages/manager/src/components/TextTooltip/TextTooltip.tsx
@@ -1,12 +1,11 @@
 import * as React from 'react';
-import ExternalLinkIcon from 'src/assets/icons/external-link.svg';
 import { makeStyles, Theme } from 'src/components/core/styles';
 import ToolTip from 'src/components/core/Tooltip';
 import Typography from 'src/components/core/Typography';
-import Link from 'src/components/Link';
 
 interface Props {
-  text: string;
+  displayText: string;
+  tooltipText: JSX.Element | string;
 }
 
 const useStyles = makeStyles((theme: Theme) => ({
@@ -29,33 +28,21 @@ const useStyles = makeStyles((theme: Theme) => ({
   },
 }));
 
-const akamaiBillingInvoiceText = (
-  <Typography>
-    Charges in the final Akamai invoice should be considered the final source
-    truth. Linode invoice will not reflect discounting, currency adjustment, or
-    any negotiated terms and conditions. Condensed and finalized invoice is
-    available within{' '}
-    <Link to="https://control.akamai.com/apps/billing">
-      Akamai Control Center &gt; Billing <ExternalLinkIcon />
-    </Link>
-    .
-  </Typography>
-);
-
-export const BillingTooltip: React.FC<Props> = (props) => {
+export const TextTooltip: React.FC<Props> = (props: Props) => {
   const classes = useStyles();
-  const { text } = props;
+  const { displayText, tooltipText } = props;
 
   return (
     <ToolTip
-      title={akamaiBillingInvoiceText}
+      title={tooltipText}
       placement="bottom"
+      enterTouchDelay={0}
       className={classes.root}
       classes={{ popper: classes.popper }}
     >
-      <Typography>{text}</Typography>
+      <Typography>{displayText}</Typography>
     </ToolTip>
   );
 };
 
-export default BillingTooltip;
+export default TextTooltip;
