+++ packages/manager/src/components/TextTooltip/index.ts
@@ -1,0 +1,2 @@
+import TextTooltip from './TextTooltip';
+export default TextTooltip;
